pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "queryreform"
version = "0.1.0"
description = "LLM-based Query Reformulation Toolkit"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Your Name"}]
dependencies = [
  "typer>=0.12.3",
  "pyyaml>=6.0.1",
  "openai>=1.40.0",
]

[project.optional-dependencies]
hf = ["datasets>=2.20.0"]
beir = ["beir>=2.0.0"]
dev = ["pytest>=7.4.0"]

[project.scripts]
queryreform = "queryreform.cli:app"
